+++ BigBang/graphlet/textlet.hpp
@@ -131,7 +131,7 @@ namespace WarGrey::SCADA {
 		WarGrey::SCADA::TextExtent unit_box;
 	};
 
-	private class Percentagelet : public virtual WarGrey::SCADA::Textlet, public virtual WarGrey::SCADA::IValuelet<float> {
+	private class Percentagelet : public virtual WarGrey::SCADA::Textlet, public virtual WarGrey::SCADA::IValuelet<double> {
 	public:
 		Percentagelet(Platform::String^ label = "", Platform::String^ subscript = "",
 			Microsoft::Graphics::Canvas::Text::CanvasTextFormat^ num_font = nullptr,
@@ -178,7 +178,7 @@ namespace WarGrey::SCADA {
 		void draw(Microsoft::Graphics::Canvas::CanvasDrawingSession^ ds, float x, float y, float Width, float Height) override;
 
 	protected:
-		void on_value_changed(float value) override;
+		void on_value_changed(double value) override;
 
 	private:
 		Microsoft::Graphics::Canvas::Text::CanvasTextFormat^ num_font;
